advanced_orm/requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.1.1
factory-boy==3.2.1
Faker==15.1.1
psycopg2-binary==2.9.3
python-dateutil==2.8.2
six==1.16.0
sqlparse==0.4.3
